+++ tests/integration/test_outcome.py
@@ -1918,7 +1918,7 @@ def test_span_outcomes_invalid(
     envelope = make_envelope()
     upstream.send_envelope(project_id, envelope)
 
-    outcomes = outcomes_consumer.get_outcomes(timeout=10.0)
+    outcomes = outcomes_consumer.get_outcomes(timeout=10.0, n=2)
     outcomes.sort(key=lambda o: sorted(o.items()))
 
     expected_outcomes = [
@@ -2044,7 +2044,7 @@ def test_global_rate_limit_by_namespace(
     # Send as many transactions as we can.
     send_buckets(transaction_limit, transaction_name, transaction_value, "d")
 
-    metrics = metrics_consumer.get_metrics(timeout=10, max_attempts=5)
+    metrics = metrics_consumer.get_metrics(timeout=10, n=5)
     assert len(metrics) == 5
 
     # The next request will trigger a rate limit, AFTER this request we should get 429s
@@ -2062,7 +2062,7 @@ def test_global_rate_limit_by_namespace(
     send_buckets(global_quota_remaining, session_name, session_value, "s")
 
     # Assert we didn't get ratelimited
-    metrics = metrics_consumer.get_metrics(timeout=10, max_attempts=4)
+    metrics = metrics_consumer.get_metrics(timeout=10, n=4)
     assert len(metrics) == 4
     outcomes_consumer.assert_empty()